requirements.txt
pycryptodome>=3.14.1
opencv-python-headless>=4.5.5.62
numpy>=1.22.2
stegano>=0.10.2
flask>=2.3.0
Pillow>=9.0.0
imageio-ffmpeg>=0.4.9
gunicorn>=21.2.0